requirements.txt
Django>=5.0
djangorestframework>=3.14
langchain>=0.1.0
langchain-openai>=0.1.0
openai>=1.0.0
pydantic>=2.0
psycopg2-binary>=2.9
redis>=5.0
celery>=5.3
tiktoken>=0.5